fix reset

+++ bin/reset.sh
@@ -13,13 +13,13 @@ function usage()
 }
 
 function args() {
-  bootstrap=0
+  delete_tfs=0
   arg_list=( "$@" )
   arg_count=${#arg_list[@]}
   arg_index=0
   while (( arg_index < arg_count )); do
     case "${arg_list[${arg_index}]}" in
-          "--delete-tfs") reset=1;;
+          "--delete-tfs") delete_tfs=1;;
           "--debug") set -x;;
                "-h") usage; exit;;
            "--help") usage; exit;;
@@ -48,7 +48,7 @@ for ns in $(kubectl get ns -o custom-columns=":metadata.name"); do
   done
 done
 
-if [ $delete_tfs -eq 1]; then
+if [ $delete_tfs -eq 1 ]; then
 
   for ns in $(kubectl get ns -o custom-columns=":metadata.name"); do
     for tf in $(kubectl get terraforms.infra.contrib.fluxcd.io -n $ns -o custom-columns=":metadata.name"); do
@@ -63,7 +63,7 @@ if [ $delete_tfs -eq 1]; then
   while ( true ); do
     echo "Waiting for terraform objects to be deleted"
     kubectl get terraforms.infra.contrib.fluxcd.io -A -o custom-columns=":metadata.namespace,:metadata.name" | grep -v "flux-system.*aws-dynamo-table" | grep -v "flux-system.*aws-s3-bucket" > /tmp/tf.list
-    objects="$wc -l /tmp/tf.list | awk '{print $1}'"
+    objects="$(wc -l /tmp/tf.list | awk '{print $1}')"
     if [[  $objects -eq 1 ]]; then
       break
     fi
@@ -72,14 +72,27 @@ if [ $delete_tfs -eq 1]; then
     sleep 5
   done
 
-  kubectl delete terraforms.infra.contrib.fluxcd.io -n flux-system aws-dynamo-table
-  kubectl delete terraforms.infra.contrib.fluxcd.io -n flux-system aws-s3-bucket
+  set +e
+  kubectl get terraforms.infra.contrib.fluxcd.io -n flux-system aws-dynamo-table > /dev/null 2>&1
+  present=$?
+  set -e
+  if [ $present -eq 0 ]; then
+    kubectl delete terraforms.infra.contrib.fluxcd.io -n flux-system aws-dynamo-table
+  fi
+  
+  set +e
+  kubectl get terraforms.infra.contrib.fluxcd.io -n flux-system aws-s3-bucket > /dev/null 2>&1
+  present=$?
+  set -e
+  if [ $present -eq 0 ]; then
+    kubectl delete terraforms.infra.contrib.fluxcd.io -n flux-system aws-s3-bucket
+  fi
 
   # Wait for remote state terraform objects to be deleted
   while ( true ); do
     echo "Waiting for terraform objects to be deleted"
     kubectl get terraforms.infra.contrib.fluxcd.io -A -o custom-columns=":metadata.namespace,:metadata.name" > /tmp/tf.list
-    objects="$wc -l /tmp/tf.list | awk '{print $1}'"
+    objects="$(wc -l /tmp/tf.list | awk '{print $1}')"
     if [[  $objects -eq 1 ]]; then
       break
     fi
@@ -89,17 +102,29 @@ if [ $delete_tfs -eq 1]; then
   done
 
 else
-  cluster_name=$(kubectl get cm -n flux-system cluster-config -o jsonpath='{.data.clusterName}')
-  aws s3 rm s3://${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state/$cluster_name --recursive
-  aws s3api delete-bucket --bucket ${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state
 
-  aws dynamodb delete-table --table-name ${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state
+  cluster_name=$(kubectl get cm -n flux-system cluster-config -o jsonpath='{.data.clusterName}')
+  set +e
+  aws s3 ls | grep -E "s3://${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state$" > /dev/null 2>&1
+  present=$?
+  set -e
+  if [ $present -eq 0 ]; then
+    aws s3 rm s3://${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state/$cluster_name --recursive
+    aws s3api delete-bucket --bucket ${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state
+  fi
+
+  set +e
+  aws dynamodb  list-tables | jq -r '.TableNames[]' | grep -E "^${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state$" > /dev/null 2>&1
+  present=$?
+  set -e
+  if [ $present -eq 0 ]; then
+    aaws dynamodb delete-table --table-name ${PREFIX_NAME}-ac-${AWS_ACCOUNT_ID}-${AWS_REGION}-tf-state
+  fi
+  
 fi
 
-etcd_container=$(docker ps -q --filter 'name=k8s_etcd_etcd-*' -l)
-docker exec "$etcd_container" rm -rf /var/lib/etcd/member;
-all_k8s_containers=$(docker ps -q --filter 'name=k8s_*')
-docker rm -f $all_k8s_containers
+echo "Reset Kubernetes"
+read -p "Press enter to continue" 
 
 # Wait for kubernetes to be ready
 while ( true ); do
@@ -108,6 +133,7 @@ while ( true ); do
   if [ -n "$started" ]; then
     break
   fi
+  sleep 1
 done
